+++ OUDS/Core/Components/Sources/Dialogs/Alert/OUDSAlertMessage.swift
@@ -51,6 +51,8 @@ import SwiftUI
 ///
 /// ## Rich text
 ///
+/// Only the alert message description and bullit list can use rich text.
+///
 /// Strong text can be used sparingly within alert messages to highlight key information.
 /// Rich text should use equivalent of Label / Medium / Strong token if possible, no other text styles or custom font weights should be used.
 /// Underlined text must not be used for emphasis, as it is commonly associated with links.
@@ -86,10 +88,10 @@ public struct OUDSAlertMessage: View {
 
     // MARK: Stored properties
 
-    private let text: String // TODO: #1405 - Move to TextualContent mayve?
+    private let text: String
     private let status: OUDSAlertStatus
     private let description: TextualContent?
-    private let bulletList: [String]
+    private let bulletList: [TextualContent]
     private let link: Self.Link?
     private let onClose: (() -> Void)?
 
@@ -137,6 +139,9 @@ public struct OUDSAlertMessage: View {
 
     // MARK: - Initializers
 
+    // For API signature consistency disable this warning
+    // swiftlint:disable function_default_parameter_at_end
+
     /// Creates an alert message.
     ///
     /// Use the `View/disabled(_:)` method to have component in disabled state.
@@ -175,16 +180,17 @@ public struct OUDSAlertMessage: View {
         }
         self.link = link
         self.onClose = onClose
-        self.bulletList = bulletList
+        self.bulletList = bulletList.map { .raw($0) }
     }
 
-    // swiftlint:disable function_default_parameter_at_end
     /// Creates an alert message with a rich text for the `description`.
     ///
     /// Use the `View/disabled(_:)` method to have component in disabled state.
     ///
     /// ```swift
-    ///     OUDSAlertMessage(label: "Success! Your form was submitted.", status: .positive)
+    ///     OUDSAlertMessage(label: "Success! Your form was submitted.",
+    ///                      status: .positive,
+    ///                      description: AttributedString(markdown: "This is **an important notice**") // Manage in your side errors for init
     /// ```
     ///
     /// - Parameters:
@@ -213,10 +219,97 @@ public struct OUDSAlertMessage: View {
         self.description = .attributed(description)
         self.link = link
         self.onClose = onClose
-        self.bulletList = bulletList
+        self.bulletList = bulletList.map { .raw($0) }
     }
 
-    // swiftlint:enable function_default_parameter_at_end
+    /// Creates an alert message with bulle tlsit containing rich text.
+    ///
+    /// Use the `View/disabled(_:)` method to have component in disabled state.
+    ///
+    /// ```swift
+    ///     OUDSAlertMessage(label: "Success! Your form was submitted.",
+    ///                      status: .positive,
+    ///                      bulletList: [
+    ///                         AttributedString("First thing"),
+    ///                         AttributedString(markdown: "One **more thing**"), // Manage in your side errors for init
+    ///                      ])
+    /// ```
+    ///
+    /// - Parameters:
+    ///   - label: Label displayed in the alert message. Main message that should be short, clear, and readable at a glance.
+    ///   - status: The status of the alert message. Its background color and its icon color are based on this status. There are two types of statuses (see ``OUDSAlertStatus``)
+    ///   - description: An optional supplementary text in an alert message. Use only when additional detail or guidance is needed beyond the label. It should remain
+    ///   short, clear and scannable, helping the user to understand what happened and what he can do next.
+    ///   - bulletList: An list of bullet points to be displayed in the alert message following the label or the optional `description`.
+    ///   - link: An optional link to be displayed in the alert message. It can be used to trigger an action.
+    ///   - onClose: An optional callback invoked when the close button is clicked. If `nil`, the close button is not displayed and the alert message
+    ///   remains visible until the context changes (e.g., the issue is resolved, the screen is refreshed). Otherwise, the alert message is dismissable and
+    ///   includes a close button, allowing the user to dismiss it when he has acknowledged the message.  Some alerts must remain visible to ensure
+    ///   user is aware of important information; others can be closed to reduce visual clutter.
+    public init(label: String,
+                status: OUDSAlertStatus = .positive,
+                description: String? = nil,
+                bulletList: [AttributedString],
+                link: Self.Link? = nil,
+                onClose: (() -> Void)? = nil)
+    {
+        text = label
+        if text.isEmpty {
+            OL.warning("The label for the OUDSAlertMessage must not be empty!")
+        }
+        self.status = status
+        self.description = if let description {
+            .raw(description)
+        } else {
+            nil
+        }
+        self.link = link
+        self.onClose = onClose
+        self.bulletList = bulletList.map { .attributed($0) }
+    }
+
+    /// Creates an alert message with description and bullet list containing rich texts.
+    ///
+    /// Use the `View/disabled(_:)` method to have component in disabled state.
+    ///
+    /// ```swift
+    ///     OUDSAlertMessage(label: "Success! Your form was submitted.",
+    ///                      status: .positive,
+    ///                      description: AttributedString(markdown: "This is **an important notice**") // Manage in your side errors for init
+    ///                      bulletList: [
+    ///                         AttributedString("First thing"),
+    ///                         AttributedString(markdown: "One **more thing**"), // Manage in your side errors for init
+    ///                      ])
+    /// ```
+    ///
+    /// - Parameters:
+    ///   - label: Label displayed in the alert message. Main message that should be short, clear, and readable at a glance.
+    ///   - status: The status of the alert message. Its background color and its icon color are based on this status. There are two types of statuses (see ``OUDSAlertStatus``)
+    ///   - description: An optional supplementary text in an alert message. Use only when additional detail or guidance is needed beyond the label. It should remain
+    ///   short, clear and scannable, helping the user to understand what happened and what he can do next.
+    ///   - bulletList: An list of bullet points to be displayed in the alert message following the label or the optional `description`.
+    ///   - link: An optional link to be displayed in the alert message. It can be used to trigger an action.
+    ///   - onClose: An optional callback invoked when the close button is clicked. If `nil`, the close button is not displayed and the alert message
+    ///   remains visible until the context changes (e.g., the issue is resolved, the screen is refreshed). Otherwise, the alert message is dismissable and
+    ///   includes a close button, allowing the user to dismiss it when he has acknowledged the message.  Some alerts must remain visible to ensure
+    ///   user is aware of important information; others can be closed to reduce visual clutter.
+    public init(label: String,
+                status: OUDSAlertStatus = .positive,
+                description: AttributedString,
+                bulletList: [AttributedString],
+                link: Self.Link? = nil,
+                onClose: (() -> Void)? = nil)
+    {
+        text = label
+        if text.isEmpty {
+            OL.warning("The label for the OUDSAlertMessage must not be empty!")
+        }
+        self.status = status
+        self.description = .attributed(description)
+        self.link = link
+        self.onClose = onClose
+        self.bulletList = bulletList.map { .attributed($0) }
+    }
 
     /// Creates an alert message with a localized label, looking up the key in the given bundle.
     ///
@@ -250,11 +343,13 @@ public struct OUDSAlertMessage: View {
                   onClose: onClose)
     }
 
-    // swiftlint:disable function_default_parameter_at_end
     /// Creates an alert message with a localized label, looking up the key in the given bundle, and a rich text description.
     ///
     /// ```swift
-    ///     OUDSAlertMessage(LocalizedStringKey("error_message"), bundle: Bundle.module, status: .negative)
+    ///     OUDSAlertMessage(LocalizedStringKey("error_message"),
+    ///                      bundle: Bundle.module,
+    ///                      status: .negative,
+    ///                      description: AttributedString(markdown: "This is **an important notice**")) // Manage in your side errors for init
     /// ```
     ///
     /// - Parameters:
@@ -272,6 +367,85 @@ public struct OUDSAlertMessage: View {
                 status: OUDSAlertStatus = .positive,
                 description: AttributedString,
                 bulletList: [String] = [],
+                link: Self.Link? = nil,
+                onClose: (() -> Void)? = nil)
+    {
+        self.init(label: key.resolved(tableName: tableName, bundle: bundle),
+                  status: status,
+                  description: description,
+                  bulletList: bulletList,
+                  link: link,
+                  onClose: onClose)
+    }
+
+    /// Creates an alert message with a localized label, looking up the key in the given bundle, and a rich text description.
+    /// Uses rich text for the bullet list.
+    ///
+    /// ```swift
+    ///     OUDSAlertMessage(LocalizedStringKey("error_message"),
+    ///                      bundle: Bundle.module,
+    ///                      status: .negative,
+    ///                      bulletList: [
+    ///                         AttributedString("First thing"),
+    ///                         AttributedString(markdown: "One **more thing**"), // Manage in your side errors for init
+    ///                      ])
+    /// ```
+    ///
+    /// - Parameters:
+    ///   - key: A `LocalizedStringKey` used to look up the label in the given bundle
+    ///   - tableName: The name of the `.strings` file, or `nil` for the default
+    ///   - bundle: The bundle in which to look up the localized string. Defaults to `Bundle.main`.
+    ///   - status: The status of the alert message, default set to *positive*
+    ///   - description: A supplementary text for the description as rich text
+    ///   - bulletList: An optional list of bullet points, default set to empty array
+    ///   - link: An optional link to be displayed in the alert message, default set to *nil*
+    ///   - onClose: An optional callback invoked when the close button is clicked, default set to *nil*
+    public init(_ key: LocalizedStringKey,
+                tableName: String? = nil,
+                bundle: Bundle = .main,
+                status: OUDSAlertStatus = .positive,
+                description: String? = nil,
+                bulletList: [AttributedString],
+                link: Self.Link? = nil,
+                onClose: (() -> Void)? = nil)
+    {
+        self.init(label: key.resolved(tableName: tableName, bundle: bundle),
+                  status: status,
+                  description: description,
+                  bulletList: bulletList,
+                  link: link,
+                  onClose: onClose)
+    }
+
+    /// Creates an alert message with a localized label, looking up the key in the given bundle, and a rich text description.
+    /// Uses rich text for the bullet list.
+    ///
+    /// ```swift
+    ///     OUDSAlertMessage(LocalizedStringKey("error_message"),
+    ///                      bundle: Bundle.module,
+    ///                      status: .negative,
+    ///                      description: AttributedString(markdown: "This is **an important notice**"), // Manage in your side errors for init
+    ///                      bulletList: [
+    ///                         AttributedString("First thing"),
+    ///                         AttributedString(markdown: "One **more thing**"), // Manage in your side errors for init
+    ///                      ])
+    /// ```
+    ///
+    /// - Parameters:
+    ///   - key: A `LocalizedStringKey` used to look up the label in the given bundle
+    ///   - tableName: The name of the `.strings` file, or `nil` for the default
+    ///   - bundle: The bundle in which to look up the localized string. Defaults to `Bundle.main`.
+    ///   - status: The status of the alert message, default set to *positive*
+    ///   - description: A supplementary text for the description as rich text
+    ///   - bulletList: An optional list of bullet points, default set to empty array
+    ///   - link: An optional link to be displayed in the alert message, default set to *nil*
+    ///   - onClose: An optional callback invoked when the close button is clicked, default set to *nil*
+    public init(_ key: LocalizedStringKey,
+                tableName: String? = nil,
+                bundle: Bundle = .main,
+                status: OUDSAlertStatus = .positive,
+                description: AttributedString,
+                bulletList: [AttributedString],
                 link: Self.Link? = nil,
                 onClose: (() -> Void)? = nil)
     {